Vhostmd: Move non-production code
This code is use only by tests.
