+++ pkg/downwardmetrics/vhostmd/BUILD.bazel
@@ -8,10 +8,7 @@ go_library(
     ],
     importpath = "kubevirt.io/kubevirt/pkg/downwardmetrics/vhostmd",
     visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/downwardmetrics/vhostmd/api:go_default_library",
-        "//pkg/util:go_default_library",
-    ],
+    deps = ["//pkg/downwardmetrics/vhostmd/api:go_default_library"],
 )
 
 go_test(
@@ -26,6 +23,7 @@ go_test(
     deps = [
         "//pkg/downwardmetrics/vhostmd/api:go_default_library",
         "//pkg/downwardmetrics/vhostmd/metrics:go_default_library",
+        "//pkg/util:go_default_library",
         "//staging/src/kubevirt.io/client-go/testutils:go_default_library",
         "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
