+++ pkg/downwardmetrics/vhostmd/disk_test.go
@@ -1,14 +1,20 @@
 package vhostmd
 
 import (
+	"encoding/binary"
+	"encoding/xml"
+	"fmt"
+	"io"
 	"os"
 	"path/filepath"
+	"strings"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
 	"kubevirt.io/kubevirt/pkg/downwardmetrics/vhostmd/api"
 	metricspkg "kubevirt.io/kubevirt/pkg/downwardmetrics/vhostmd/metrics"
+	"kubevirt.io/kubevirt/pkg/util"
 )
 
 var _ = Describe("vhostmd", func() {
@@ -86,5 +92,75 @@ var _ = Describe("vhostmd", func() {
 			}
 		})
 	})
-
 })
+
+func (d *Disk) Metrics() (*api.Metrics, error) {
+	m := &api.Metrics{}
+	if err := xml.Unmarshal(d.Raw, m); err != nil {
+		return nil, err
+	}
+	m.Text = strings.TrimSpace(m.Text)
+	for i, metric := range m.Metrics {
+		m.Metrics[i].Name = strings.TrimSpace(metric.Name)
+		m.Metrics[i].Type = api.MetricType(strings.TrimSpace(string(metric.Type)))
+		m.Metrics[i].Context = api.MetricContext(strings.TrimSpace(string(metric.Context)))
+		m.Metrics[i].Value = strings.TrimSpace(metric.Value)
+		m.Metrics[i].Text = strings.TrimSpace(metric.Text)
+	}
+	return m, nil
+}
+
+func (d *Disk) Verify() error {
+	var checksum int32
+	for _, b := range d.Raw {
+		checksum = checksum + int32(b)
+	}
+	if d.Header.Flag > 0 {
+		return fmt.Errorf("file is locked")
+	}
+	if checksum != d.Header.Checksum {
+		return fmt.Errorf("checksum is %v, but expected %v", checksum, d.Header.Checksum)
+	}
+	return nil
+}
+
+func (v *vhostmd) Read() (*api.Metrics, error) {
+	disk, err := readDisk(v.filePath)
+	if err != nil {
+		return nil, fmt.Errorf("failed to load vhostmd file: %v", err)
+	}
+	if err := disk.Verify(); err != nil {
+		return nil, fmt.Errorf("failed to verify vhostmd file: %v", err)
+	}
+	return disk.Metrics()
+}
+
+func readDisk(filePath string) (*Disk, error) {
+	f, err := os.Open(filePath)
+	if err != nil {
+		return nil, err
+	}
+	// If the read operation succeeds, but close fails, we have already read the data,
+	// so it is ok to not return the error.
+	defer util.CloseIOAndCheckErr(f, nil)
+
+	d := &Disk{
+		Header: &Header{},
+	}
+	if err = binary.Read(f, binary.BigEndian, d.Header); err != nil {
+		return nil, err
+	}
+
+	if d.Header.Flag == 0 {
+		if d.Header.Length > maxBodyLength {
+			return nil, fmt.Errorf("Invalid metrics file. Expected a maximum body length of %v, got %v", maxBodyLength, d.Header.Length)
+		}
+
+		d.Raw = make([]byte, d.Header.Length)
+
+		if _, err = io.ReadFull(f, d.Raw); err != nil {
+			return nil, err
+		}
+	}
+	return d, err
+}
